\name{EpiTOCcpgs}
\alias{EpiTOCcpgs}
\docType{data}
\title{
List of CpGs used in EpiTOC algorithm.
}
\description{
CpGs for EpiTOC algorithm described in Yang et al 2016.
}
\usage{data("EpiTOCcpgs")}
\format{
  A character vector of CpG IDs as character strings.
}
\details{
Yang et al. "Correlation of an epigenetic mitotic clock with cancer risk." Genome Biology, 2016.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(EpiTOCcpgs)
}
\keyword{datasets}
